change name HomeScreen to Home
