+++ src/screens/HomeScreen/Home.js
@@ -1,18 +1,71 @@
-import React from 'react'
-import { Text, View } from 'react-native'
-import { StyleSheet } from 'react-native';
+import React from 'react';
+import {
+  View,
+  Text,
+  Image,
+  StyleSheet,
+  StatusBar,
+  TouchableOpacity,
+  ScrollView,
+} from 'react-native';
+import styles from './styles';
+import { Constants } from 'expo';
 
-export default function Home() {
-    return (
-        <View style={styles.container}>
-          <Text>This Home screen</Text>
+import {useTheme} from '@react-navigation/native';
+import Ionicons from 'react-native-vector-icons/Ionicons';
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import Fontisto from 'react-native-vector-icons/Fontisto';
+import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
+
+const Home = ({navigation}) => {
+    const theme = useTheme();
+
+return (
+    <View style={styles.cardsWrapper}>
+            <Text>POST A JOB</Text>
+
+            <Text
+            style={{
+                alignSelf: 'center',
+                fontSize: 18,
+                fontWeight: 'bold',
+                color: '#333',
+            }}>
+            MY POSTED JOBS
+            </Text>
+            
+            <View style={styles.container}>
+                <View style={styles.rectangle}></View>
+                {/* <Image source={require('assets/car1.png')} /> */}
+            </View>
+
+        <View style={styles.card}>
+            <View style={styles.cardImgWrapper}>
+                
+                <Text style={styles.cardTitle}>ADVERT HEADING</Text>
+            </View>
+
+            <View style={styles.cardImgWrapper}>
+                <Image
+                //source={require('../assets/car2.png')}
+                resizeMode="cover"
+                style={styles.cardImg}
+                />
+                <Text style={styles.cardTitle}>ADVERT HEADING</Text>
+            </View>
+
+            <View style={styles.cardImgWrapper}>
+                <Image
+                //source={require('../assets/car3.png')}
+                resizeMode="cover"
+                style={styles.cardImg}
+                />
+                <Text style={styles.cardTitle}>ADVERT HEADING</Text>
+            </View>
+            
         </View>
+    </View>
     )
-}
+};
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: 'center'
-    },
-})
+export default Home;
