+++ src/screens/index.js
@@ -2,5 +2,5 @@ export { default as ServiceScreen } from './ServiceScreen/ServiceScreen';
 
 export { default as NavigationScreen } from './NavigationScreen/NavigationScreen';
 
-export { default as HomeScreen } from './HomeScreen/HomeScreen';
+export { default as Home } from './HomeScreen/Home';
 export { default as AddItemScreen } from './AddItemScreen/AddItemScreen';
